+++ nengo_spa/modules/tests/test_bind.py
@@ -72,5 +72,5 @@ def test_unbind(Simulator, algebra, side, seed, rng):
         sim.run(0.2)
 
     assert_sp_close(
-        sim.trange(), sim.data[p], vocab.parse("A * B * ~B"), skip=0.15, atol=0.3
+        sim.trange(), sim.data[p], vocab.parse("A * B * B.rinv()"), skip=0.15, atol=0.3
     )